posix: Do not refute listings including setuid/setgid/sticky bits
As written, the regex that matches on POSIX directory listings misses
some cases that can appear, especially on more complex systems.  These
are cases where the set-user-ID on execution, set-group-ID on execution,
and saved-text (aka setuid, setgid, and "sticky" bits respectively) are
set, typically by a system administrator.

These are quite consistently represented by 's' and 't' in place of the
eXecute bit, and the uppercase alternatives ('S' and 'T') if the eXecute
bit is not set, but the corresponding setuid/setgid/sticky bit is set.

Given that the extant regex does not attempt to validate the ordering
(i.e., xwrxwrxwr is entirely valid), I chose to simply extend the set of
valid values that are accepted in the string of 9 characters following
the initial -/l/d part.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>