fix(async): Fixed `retr` method signature on the `AsyncFtpStream` to allow passing a closure taking the stream reader. (#110)
The signature of the `retr` method was not allowing any argument, because the dyn Read needs to be Unpin, but it's a mutable reference at the same time and this with Async causes several issues. The mutable reference is required by the finalize_retr_stream which is called immediately after calling the closure. Because of this, the signature has been changed to return both the result and the stream back to be able to finalize it.

closes #88